Remove the navigation error and fix all the bugs and also add the App Storage.
