+++ softo-team.xcodeproj/xcuserdata/user.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "BA1AA95C-027D-4BE0-9467-76ED3266310D"
+   type = "1"
+   version = "2.0">
+</Bucket>
