+++ softo-team/Onboardings/Onboarding3.swift
@@ -10,17 +10,19 @@ import Firebase
 
 struct Onboarding3: View {
     @State private var progress: CGFloat = 3.0 / 4.0 // Update progress for Onboarding 3
-    @State private var emergencyContactName: String = ""
-    @State private var emergencyContactNumber: String = ""
-    @State private var emergencyContactName1: String = ""
-    @State private var emergencyContactNumber1: String = ""
-    @State private var selectedCountryCode = "+1"
+    @AppStorage("emergencyContactName") private var emergencyContactName: String = ""
+    @AppStorage("emergencyContactNumber") private var emergencyContactNumber: String = ""
+    @AppStorage("emergencyContactName1") private var emergencyContactName1: String = ""
+    @AppStorage("emergencyContactNumber1") private var emergencyContactNumber1: String = ""
+    @AppStorage("selectedCountryCode") private var selectedCountryCode: String = "+1"
+    @AppStorage("selectedCountryCode1") private var selectedCountryCode1: String = "+1"
     @State private var isOnboarding4Active = false
     @State private var showAlert = false
     @State private var countryCodes: [String] = [
         "+1", "+7", "+20", "+27", "+30", "+31", "+32", "+33", "+34", "+36", "+39", "+40", "+41", "+43", "+44", "+45", "+46", "+47", "+48", "+49", "+51", "+52", "+53", "+54", "+55", "+56", "+57", "+58", "+60", "+61", "+62", "+63", "+64", "+65", "+66", "+81", "+82", "+84", "+86", "+90", "+91", "+92", "+93", "+94", "+95", "+98", "+212", "+213", "+216", "+218", "+220", "+221", "+222", "+223", "+224", "+225", "+226", "+227", "+228", "+229", "+230", "+231", "+232", "+233", "+234", "+235", "+236", "+237", "+238", "+239", "+240", "+241", "+242", "+243", "+244", "+245", "+246", "+248", "+249", "+250", "+251", "+252", "+253", "+254", "+255", "+256", "+257", "+258", "+260", "+261", "+262", "+263", "+264", "+265", "+266", "+267", "+268", "+269", "+290", "+291", "+297", "+298", "+299", "+350", "+351", "+352", "+353", "+354", "+355", "+356", "+357", "+358", "+359", "+370", "+371", "+372", "+373", "+374", "+375", "+376", "+377", "+378", "+379", "+380", "+381", "+382", "+383", "+385", "+386", "+387", "+389", "+420", "+421", "+423", "+500", "+501", "+502", "+503", "+504", "+505", "+506", "+507", "+508", "+509", "+590", "+591", "+592", "+593", "+594", "+595", "+596", "+597", "+598", "+599", "+670", "+672", "+673", "+674", "+675", "+676", "+677", "+678", "+679", "+680", "+681", "+682", "+683", "+685", "+686", "+687", "+688", "+689", "+690", "+691", "+692", "+850", "+852", "+853", "+855", "+856", "+880", "+886", "+960", "+961", "+962", "+963", "+964", "+965", "+966", "+967", "+968", "+970", "+971", "+972", "+973", "+974", "+975", "+976", "+977", "+992", "+993", "+994", "+995", "+996", "+998", "+1242", "+1246", "+1264", "+1268", "+1284", "+1340", "+1345", "+1441", "+1473", "+1649", "+1664", "+1670", "+1671", "+1684", "+1721", "+1758", "+1767", "+1784", "+1787", "+1809", "+1829", "+1849", "+1868", "+1869", "+1876", "+1939"] //  country codes
 
     var body: some View {
+        ScrollView{
         VStack {
 
             // Title
@@ -41,26 +43,26 @@ struct Onboarding3: View {
             .padding()
             
             
-            // Form for Emergency Contact
-            Form {
-                Section(header: Text("Enter Emergency Contact Information")) {
-                    // Emergency Contact 1
-                    VStack(alignment: .leading) {
-                        Text("Emergency Contact 1")
-                            .font(.headline)
-                            .padding()
-                        
-                        TextField("Contact Name", text: $emergencyContactName)
-                            .padding()
-                            .textFieldStyle(.roundedBorder)
-                        HStack{
-                            Picker("Country Code", selection: $selectedCountryCode) {
-                                ForEach(countryCodes, id: \.self) {
-                                    Text($0)
-                                }
+            VStack(alignment: .leading, spacing: 20) {
+                // Emergency Contact 1
+                VStack(alignment: .leading) {
+                    Text("Emergency Contact 1")
+                        .font(.headline)
+                        .padding()
+
+                    TextField("Contact Name", text: $emergencyContactName)
+                        .padding()
+                        .textFieldStyle(.roundedBorder)
+
+                    HStack {
+                        Picker("Country Code", selection: $selectedCountryCode) {
+                            ForEach(countryCodes, id: \.self) {
+                                Text($0)
                             }
-                            .pickerStyle(MenuPickerStyle())
-                            .frame(width: 20)
+                        }
+                        .pickerStyle(MenuPickerStyle())
+                        .frame(width: 20)
+
                         TextField("eg:3000000000", text: $emergencyContactNumber)
                             .padding()
                             .textFieldStyle(.roundedBorder)
@@ -73,8 +75,10 @@ struct Onboarding3: View {
                                     emergencyContactNumber = filtered
                                 }
                             }
-                        }
                     }
+                }
+            }
+
                     .padding()
                     
                     // Emergency Contact 2
@@ -87,7 +91,7 @@ struct Onboarding3: View {
                             .padding()
                             .textFieldStyle(.roundedBorder)
                         HStack{
-                            Picker("Country Code", selection: $selectedCountryCode) {
+                            Picker("Country Code", selection: $selectedCountryCode1) {
                                 ForEach(countryCodes, id: \.self) {
                                     Text($0)
                                 }
@@ -112,8 +116,7 @@ struct Onboarding3: View {
                     .padding()
                     }
                 }
-            }
-
+            
             // Button
             Spacer().frame(height: 50)
             // NavigationLink (Next)
@@ -134,7 +137,12 @@ struct Onboarding3: View {
             }
             .buttonStyle(.borderedProminent)
             .padding()
-        }
+            .onAppear(){
+                emergencyContactName=""
+                emergencyContactNumber=""
+                emergencyContactName1=""
+                emergencyContactNumber1=""
+            }
         .alert(isPresented: $showAlert) {
             Alert(
                 title: Text("Error"),
@@ -143,7 +151,7 @@ struct Onboarding3: View {
             )
         }
     }
-
+    }
     // Function to check if all fields are filled
     private func allFieldsAreFilled() -> Bool {
         return !emergencyContactName.isEmpty && !emergencyContactNumber.isEmpty
@@ -160,9 +168,9 @@ struct Onboarding3: View {
             // Update user data with emergency contact information
             let emergencyContactData: [String: Any] = [
                 "emergencyContactName": emergencyContactName,
-                "emergencyContactNumber": emergencyContactNumber,
+                "emergencyContactNumber": "\(selectedCountryCode) \(emergencyContactNumber)",
                 "emergencyContactName1": emergencyContactName1,
-                "emergencyContactNumber1": emergencyContactNumber1
+                "emergencyContactNumber1": "\(selectedCountryCode1) \(emergencyContactNumber)",
                 // Add other fields specific to emergency contact if needed
             ]
 
@@ -177,9 +185,7 @@ struct Onboarding3: View {
             }
         }
     }
-
 }
-
 struct Onboarding3_Previews: PreviewProvider {
     static var previews: some View {
         Onboarding3()
